Move scalarvalue
