+++ datafusion/proto/proto/datafusion.proto
@@ -327,7 +327,7 @@ message LogicalExprNode {
     // alias
     AliasNode alias = 2;
 
-    ScalarValue literal = 3;
+    datafusion_common.ScalarValue literal = 3;
 
     // binary expressions
     BinaryExprNode binary_expr = 4;
@@ -410,7 +410,7 @@ message RollupNode {
 }
 
 message NamedStructField {
-  ScalarValue name = 1;
+  datafusion_common.ScalarValue name = 1;
 }
 
 message ListIndex {
@@ -666,7 +666,7 @@ enum WindowFrameBoundType {
 
 message WindowFrameBound {
   WindowFrameBoundType window_frame_bound_type = 1;
-  ScalarValue bound_value = 2;
+  datafusion_common.ScalarValue bound_value = 2;
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
@@ -680,138 +680,6 @@ message FixedSizeBinary{
 enum DateUnit{
   Day = 0;
   DateMillisecond = 1;
-}
-
-// Used for List/FixedSizeList/LargeList/Struct
-message ScalarNestedValue {
-  message Dictionary {
-    bytes ipc_message = 1;
-    bytes arrow_data = 2;
-  }
-
-  bytes ipc_message = 1;
-  bytes arrow_data = 2;
-  datafusion_common.Schema schema = 3;
-  repeated Dictionary dictionaries = 4;
-}
-
-message ScalarTime32Value {
-  oneof value {
-    int32 time32_second_value = 1;
-    int32 time32_millisecond_value = 2;
-  };
-}
-
-message ScalarTime64Value {
-  oneof value {
-    int64 time64_microsecond_value = 1;
-    int64 time64_nanosecond_value = 2;
-  };
-}
-
-message ScalarTimestampValue {
-  oneof value {
-    int64 time_microsecond_value = 1;
-    int64 time_nanosecond_value = 2;
-    int64 time_second_value = 3;
-    int64 time_millisecond_value = 4;
-  };
-  string timezone = 5;
-}
-
-message ScalarDictionaryValue {
-  datafusion_common.ArrowType index_type = 1;
-  ScalarValue value = 2;
-}
-
-message IntervalMonthDayNanoValue {
-  int32 months = 1;
-  int32 days = 2;
-  int64 nanos = 3;
-}
-
-message UnionField {
-  int32 field_id = 1;
-  datafusion_common.Field field = 2;
-}
-
-message UnionValue {
-  // Note that a null union value must have one or more fields, so we
-  // encode a null UnionValue as one with value_id == 128
-  int32 value_id = 1;
-  ScalarValue value = 2;
-  repeated UnionField fields = 3;
-  datafusion_common.UnionMode mode = 4;
-}
-
-message ScalarFixedSizeBinary{
-  bytes values = 1;
-  int32 length = 2;
-}
-
-message ScalarValue{
-  // was PrimitiveScalarType null_value = 19;
-  reserved 19;
-
-  oneof value {
-    // was PrimitiveScalarType null_value = 19;
-    // Null value of any type
-    datafusion_common.ArrowType null_value = 33;
-
-    bool   bool_value = 1;
-    string utf8_value = 2;
-    string large_utf8_value = 3;
-    int32  int8_value = 4;
-    int32  int16_value = 5;
-    int32  int32_value = 6;
-    int64  int64_value = 7;
-    uint32 uint8_value = 8;
-    uint32 uint16_value = 9;
-    uint32 uint32_value = 10;
-    uint64 uint64_value = 11;
-    float  float32_value = 12;
-    double float64_value = 13;
-    // Literal Date32 value always has a unit of day
-    int32  date_32_value = 14;
-    ScalarTime32Value time32_value = 15;
-    ScalarNestedValue large_list_value = 16;
-    ScalarNestedValue list_value = 17;
-    ScalarNestedValue fixed_size_list_value = 18;
-    ScalarNestedValue struct_value = 32;
-
-    Decimal128 decimal128_value = 20;
-    Decimal256 decimal256_value = 39;
-
-    int64 date_64_value = 21;
-    int32 interval_yearmonth_value = 24;
-    int64 interval_daytime_value = 25;
-
-    int64 duration_second_value = 35;
-    int64 duration_millisecond_value = 36;
-    int64 duration_microsecond_value = 37;
-    int64 duration_nanosecond_value = 38;
-
-    ScalarTimestampValue timestamp_value = 26;
-    ScalarDictionaryValue dictionary_value = 27;
-    bytes binary_value = 28;
-    bytes large_binary_value = 29;
-    ScalarTime64Value time64_value = 30;
-    IntervalMonthDayNanoValue interval_month_day_nano = 31;
-    ScalarFixedSizeBinary fixed_size_binary_value = 34;
-    UnionValue union_value = 42;
-  }
-}
-
-message Decimal128{
-  bytes value = 1;
-  int64 p = 2;
-  int64 s = 3;
-}
-
-message Decimal256{
-  bytes value = 1;
-  int64 p = 2;
-  int64 s = 3;
 }
 
 message AnalyzedLogicalPlanType {
@@ -1134,7 +1002,7 @@ message PhysicalExprNode {
     // column references
     PhysicalColumn column = 1;
 
-    ScalarValue literal = 2;
+    datafusion_common.ScalarValue literal = 2;
 
     // binary expressions
     PhysicalBinaryExprNode binary_expr = 3;
@@ -1531,7 +1399,7 @@ message PartitionedFile {
   string path = 1;
   uint64 size = 2;
   uint64 last_modified_ns = 3;
-  repeated ScalarValue partition_values = 4;
+  repeated datafusion_common.ScalarValue partition_values = 4;
   FileRange range = 5;
   Statistics statistics = 6;
 }
@@ -1550,7 +1418,7 @@ message PartitionStats {
 
 message Precision{
   PrecisionInfo precision_info = 1;
-  ScalarValue val = 2;
+  datafusion_common.ScalarValue val = 2;
 }
 
 enum PrecisionInfo {
